Fix: switch fp16 to bf16 for A40/Ampere GPU compatibility
A40 uses BFloat16 natively. fp16 gradient scaler throws
NotImplementedError on Ampere GPUs. bf16 is also more
training-stable than fp16.

+++ train_sentiment.py
@@ -80,7 +80,7 @@ def train() -> None:
         save_strategy="epoch",
         dataset_text_field="text",
         max_length=512,
-        fp16=(device == "cuda"),
+        bf16=(device == "cuda"),
         report_to="none",
     )
